fix biological replicate, fraction identifer, technical replicate
No fractionation of samples was mentioned in metadata or manuscript so setting all values in fraction column to 1.

changed characteristics[technical replicate] to comment[technical replicate] and removed duplicate technical replicate column.
Set technical replicates to all 1 as there was no technical replication - each of the 24 samples run in this dataset were a different biopsy so biological replicates must be set, not technical